rename files registration -> register
